% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air-package.R
\docType{package}
\name{air-package}
\alias{air}
\alias{air-package}
\title{air: R Interface to the 'Air' Formatter}
\description{
Format R files using the 'Air' formatter.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/air-r}
  \item Report bugs at \url{https://github.com/DavisVaughan/air-r/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@posit.co}

Authors:
\itemize{
  \item Lionel Henry \email{lionel@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
